优化了启动逻辑
